refactor(ast): extract visit APIs into a fn
BREAKING CHANGE: ast nodes will no longer have .accept methods
