+++ packages/runtime/src/binding/ast.ts
@@ -3,7 +3,7 @@ import { type IBinding } from '../observation';
 import { Scope } from '../observation/binding-context';
 import { ISignaler } from '../observation/signaler';
 import { type IConnectableBinding } from './connectable';
-import { safeString, isArray, isFunction, isString } from '../utilities-objects';
+import { safeString, isArray, isFunction } from '../utilities-objects';
 
 import type { IIndexable, IServiceLocator } from '@aurelia/kernel';
 import type {
@@ -12,6 +12,12 @@ import type {
   IConnectable,
   ISubscriber,
 } from '../observation';
+
+export {
+  IVisitor,
+  Unparser,
+  visitAst,
+} from './ast.visitor';
 
 export const enum ExpressionKind {
   AccessThis,
@@ -65,351 +71,6 @@ export type AnyBindingExpression = Interpolation | ForOfStatement | IsBindingBeh
 export interface IExpressionHydrator {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   hydrate(jsonExpr: any): any;
-}
-
-export interface IVisitor<T = unknown> {
-  visitAccessKeyed(expr: AccessKeyedExpression): T;
-  visitAccessMember(expr: AccessMemberExpression): T;
-  visitAccessScope(expr: AccessScopeExpression): T;
-  visitAccessThis(expr: AccessThisExpression): T;
-  visitArrayBindingPattern(expr: ArrayBindingPattern): T;
-  visitArrayLiteral(expr: ArrayLiteralExpression): T;
-  visitArrowFunction(expr: ArrowFunction): T;
-  visitAssign(expr: AssignExpression): T;
-  visitBinary(expr: BinaryExpression): T;
-  visitBindingBehavior(expr: BindingBehaviorExpression): T;
-  visitBindingIdentifier(expr: BindingIdentifier): T;
-  visitCallFunction(expr: CallFunctionExpression): T;
-  visitCallMember(expr: CallMemberExpression): T;
-  visitCallScope(expr: CallScopeExpression): T;
-  visitConditional(expr: ConditionalExpression): T;
-  visitForOfStatement(expr: ForOfStatement): T;
-  visitInterpolation(expr: Interpolation): T;
-  visitObjectBindingPattern(expr: ObjectBindingPattern): T;
-  visitObjectLiteral(expr: ObjectLiteralExpression): T;
-  visitPrimitiveLiteral(expr: PrimitiveLiteralExpression): T;
-  visitTaggedTemplate(expr: TaggedTemplateExpression): T;
-  visitTemplate(expr: TemplateExpression): T;
-  visitUnary(expr: UnaryExpression): T;
-  visitValueConverter(expr: ValueConverterExpression): T;
-  visitDestructuringAssignmentExpression(expr: DestructuringAssignmentExpression): T;
-  visitDestructuringAssignmentSingleExpression(expr: DestructuringAssignmentSingleExpression): T;
-  visitDestructuringAssignmentRestExpression(expr: DestructuringAssignmentRestExpression): T;
-}
-
-export class Unparser implements IVisitor<void> {
-  public text: string = '';
-
-  public static unparse(expr: IsExpressionOrStatement): string {
-    const visitor = new Unparser();
-    expr.accept(visitor);
-    return visitor.text;
-  }
-
-  public visitAccessMember(expr: AccessMemberExpression): void {
-    expr.object.accept(this);
-    this.text += `${expr.optional ? '?' : ''}.${expr.name}`;
-  }
-
-  public visitAccessKeyed(expr: AccessKeyedExpression): void {
-    expr.object.accept(this);
-    this.text += `${expr.optional ? '?.' : ''}[`;
-    expr.key.accept(this);
-    this.text += ']';
-  }
-
-  public visitAccessThis(expr: AccessThisExpression): void {
-    if (expr.ancestor === 0) {
-      this.text += '$this';
-      return;
-    }
-    this.text += '$parent';
-    let i = expr.ancestor - 1;
-    while (i--) {
-      this.text += '.$parent';
-    }
-  }
-
-  public visitAccessScope(expr: AccessScopeExpression): void {
-    let i = expr.ancestor;
-    while (i--) {
-      this.text += '$parent.';
-    }
-    this.text += expr.name;
-  }
-
-  public visitArrayLiteral(expr: ArrayLiteralExpression): void {
-    const elements = expr.elements;
-    this.text += '[';
-    for (let i = 0, length = elements.length; i < length; ++i) {
-      if (i !== 0) {
-        this.text += ',';
-      }
-      elements[i].accept(this);
-    }
-    this.text += ']';
-  }
-
-  public visitArrowFunction(expr: ArrowFunction): void {
-    const args = expr.args;
-    const ii = args.length;
-    let i = 0;
-    let text = '(';
-    let name: string;
-    for (; i < ii; ++i) {
-      name = args[i].name;
-      if (i > 0) {
-        text += ', ';
-      }
-      if (i < ii - 1) {
-        text += name;
-      } else {
-        text += expr.rest ? `...${name}` : name;
-      }
-    }
-    this.text += `${text}) => `;
-    expr.body.accept(this);
-  }
-
-  public visitObjectLiteral(expr: ObjectLiteralExpression): void {
-    const keys = expr.keys;
-    const values = expr.values;
-    this.text += '{';
-    for (let i = 0, length = keys.length; i < length; ++i) {
-      if (i !== 0) {
-        this.text += ',';
-      }
-      this.text += `'${keys[i]}':`;
-      values[i].accept(this);
-    }
-    this.text += '}';
-  }
-
-  public visitPrimitiveLiteral(expr: PrimitiveLiteralExpression): void {
-    this.text += '(';
-    if (isString(expr.value)) {
-      const escaped = expr.value.replace(/'/g, '\\\'');
-      this.text += `'${escaped}'`;
-    } else {
-      this.text += `${expr.value}`;
-    }
-    this.text += ')';
-  }
-
-  public visitCallFunction(expr: CallFunctionExpression): void {
-    this.text += '(';
-    expr.func.accept(this);
-    this.text += expr.optional ? '?.' : '';
-    this.writeArgs(expr.args);
-    this.text += ')';
-  }
-
-  public visitCallMember(expr: CallMemberExpression): void {
-    this.text += '(';
-    expr.object.accept(this);
-    this.text += `${expr.optionalMember ? '?.' : ''}.${expr.name}${expr.optionalCall ? '?.' : ''}`;
-    this.writeArgs(expr.args);
-    this.text += ')';
-  }
-
-  public visitCallScope(expr: CallScopeExpression): void {
-    this.text += '(';
-    let i = expr.ancestor;
-    while (i--) {
-      this.text += '$parent.';
-    }
-    this.text += `${expr.name}${expr.optional ? '?.' : ''}`;
-    this.writeArgs(expr.args);
-    this.text += ')';
-  }
-
-  public visitTemplate(expr: TemplateExpression): void {
-    const { cooked, expressions } = expr;
-    const length = expressions.length;
-    this.text += '`';
-    this.text += cooked[0];
-    for (let i = 0; i < length; i++) {
-      expressions[i].accept(this);
-      this.text += cooked[i + 1];
-    }
-    this.text += '`';
-  }
-
-  public visitTaggedTemplate(expr: TaggedTemplateExpression): void {
-    const { cooked, expressions } = expr;
-    const length = expressions.length;
-    expr.func.accept(this);
-    this.text += '`';
-    this.text += cooked[0];
-    for (let i = 0; i < length; i++) {
-      expressions[i].accept(this);
-      this.text += cooked[i + 1];
-    }
-    this.text += '`';
-  }
-
-  public visitUnary(expr: UnaryExpression): void {
-    this.text += `(${expr.operation}`;
-    if (expr.operation.charCodeAt(0) >= /* a */97) {
-      this.text += ' ';
-    }
-    expr.expression.accept(this);
-    this.text += ')';
-  }
-
-  public visitBinary(expr: BinaryExpression): void {
-    this.text += '(';
-    expr.left.accept(this);
-    if (expr.operation.charCodeAt(0) === /* i */105) {
-      this.text += ` ${expr.operation} `;
-    } else {
-      this.text += expr.operation;
-    }
-    expr.right.accept(this);
-    this.text += ')';
-  }
-
-  public visitConditional(expr: ConditionalExpression): void {
-    this.text += '(';
-    expr.condition.accept(this);
-    this.text += '?';
-    expr.yes.accept(this);
-    this.text += ':';
-    expr.no.accept(this);
-    this.text += ')';
-  }
-
-  public visitAssign(expr: AssignExpression): void {
-    this.text += '(';
-    expr.target.accept(this);
-    this.text += '=';
-    expr.value.accept(this);
-    this.text += ')';
-  }
-
-  public visitValueConverter(expr: ValueConverterExpression): void {
-    const args = expr.args;
-    expr.expression.accept(this);
-    this.text += `|${expr.name}`;
-    for (let i = 0, length = args.length; i < length; ++i) {
-      this.text += ':';
-      args[i].accept(this);
-    }
-  }
-
-  public visitBindingBehavior(expr: BindingBehaviorExpression): void {
-    const args = expr.args;
-    expr.expression.accept(this);
-    this.text += `&${expr.name}`;
-    for (let i = 0, length = args.length; i < length; ++i) {
-      this.text += ':';
-      args[i].accept(this);
-    }
-  }
-
-  public visitArrayBindingPattern(expr: ArrayBindingPattern): void {
-    const elements = expr.elements;
-    this.text += '[';
-    for (let i = 0, length = elements.length; i < length; ++i) {
-      if (i !== 0) {
-        this.text += ',';
-      }
-      elements[i].accept(this);
-    }
-    this.text += ']';
-  }
-
-  public visitObjectBindingPattern(expr: ObjectBindingPattern): void {
-    const keys = expr.keys;
-    const values = expr.values;
-    this.text += '{';
-    for (let i = 0, length = keys.length; i < length; ++i) {
-      if (i !== 0) {
-        this.text += ',';
-      }
-      this.text += `'${keys[i]}':`;
-      values[i].accept(this);
-    }
-    this.text += '}';
-  }
-
-  public visitBindingIdentifier(expr: BindingIdentifier): void {
-    this.text += expr.name;
-  }
-
-  public visitForOfStatement(expr: ForOfStatement): void {
-    expr.declaration.accept(this);
-    this.text += ' of ';
-    expr.iterable.accept(this);
-  }
-
-  public visitInterpolation(expr: Interpolation): void {
-    const { parts, expressions } = expr;
-    const length = expressions.length;
-    this.text += '${';
-    this.text += parts[0];
-    for (let i = 0; i < length; i++) {
-      expressions[i].accept(this);
-      this.text += parts[i + 1];
-    }
-    this.text += '}';
-  }
-
-  public visitDestructuringAssignmentExpression(expr: DestructuringAssignmentExpression): void {
-    const $kind = expr.$kind;
-    const isObjDes = $kind === ExpressionKind.ObjectDestructuring;
-    this.text += isObjDes ? '{' : '[';
-    const list = expr.list;
-    const len = list.length;
-    let i: number;
-    let item: DestructuringAssignmentExpression | DestructuringAssignmentSingleExpression | DestructuringAssignmentRestExpression;
-    for(i = 0; i< len; i++) {
-      item = list[i];
-      switch(item.$kind) {
-        case ExpressionKind.DestructuringAssignmentLeaf:
-          item.accept(this);
-          break;
-        case ExpressionKind.ArrayDestructuring:
-        case ExpressionKind.ObjectDestructuring: {
-          const source = item.source;
-          if(source) {
-            source.accept(this);
-            this.text += ':';
-          }
-          item.accept(this);
-          break;
-        }
-      }
-    }
-    this.text += isObjDes ? '}' : ']';
-  }
-
-  public visitDestructuringAssignmentSingleExpression(expr: DestructuringAssignmentSingleExpression): void {
-    expr.source.accept(this);
-    this.text += ':';
-    expr.target.accept(this);
-    const initializer = expr.initializer;
-    if(initializer !== void 0) {
-      this.text +='=';
-      initializer.accept(this);
-    }
-  }
-
-  public visitDestructuringAssignmentRestExpression(expr: DestructuringAssignmentRestExpression): void {
-    this.text += '...';
-    expr.accept(this);
-  }
-
-  private writeArgs(args: readonly IsBindingBehavior[]): void {
-    this.text += '(';
-    for (let i = 0, length = args.length; i < length; ++i) {
-      if (i !== 0) {
-        this.text += ',';
-      }
-      args[i].accept(this);
-    }
-    this.text += ')';
-  }
 }
 
 /**
@@ -504,14 +165,6 @@ export class BindingBehaviorExpression {
       this.expression.unbind(s, b);
     }
   }
-
-  public accept<T>(visitor: IVisitor<T>): T {
-    return visitor.visitBindingBehavior(this);
-  }
-
-  public toString(): string {
-    return Unparser.unparse(this);
-  }
 }
 
 const behaviorNotFoundError = (name: string) =>
@@ -603,14 +256,6 @@ export class ValueConverterExpression {
       signaler.removeSignalListener(vc.signals[i], b as unknown as ISubscriber);
     }
   }
-
-  public accept<T>(visitor: IVisitor<T>): T {
-    return visitor.visitValueConverter(this);
-  }
-
-  public toString(): string {
-    return Unparser.unparse(this);
-  }
 }
 
 const converterNotFoundError = (name: string) => {
@@ -638,14 +283,6 @@ export class AssignExpression {
     this.value.assign(s, e, val);
     return this.target.assign(s, e, val);
   }
-
-  public accept<T>(visitor: IVisitor<T>): T {
-    return visitor.visitAssign(this);
-  }
-
-  public toString(): string {
-    return Unparser.unparse(this);
-  }
 }
 
 export class ConditionalExpression {
@@ -666,14 +303,6 @@ export class ConditionalExpression {
 
   public assign(_s: Scope, _e: IAstEvaluator | null, _obj: unknown): unknown {
     return void 0;
-  }
-
-  public accept<T>(visitor: IVisitor<T>): T {
-    return visitor.visitConditional(this);
-  }
-
-  public toString(): string {
-    return Unparser.unparse(this);
   }
 }
 
@@ -699,14 +328,6 @@ export class AccessThisExpression {
 
   public assign(_s: Scope, _e: IAstEvaluator | null, _obj: unknown): unknown {
     return void 0;
-  }
-
-  public accept<T>(visitor: IVisitor<T>): T {
-    return visitor.visitAccessThis(this);
-  }
-
-  public toString(): string {
-    return Unparser.unparse(this);
   }
 }
 
@@ -761,14 +382,6 @@ export class AccessScopeExpression {
         return obj[this.name] = val;
       }
     }
-  }
-
-  public accept<T>(visitor: IVisitor<T>): T {
-    return visitor.visitAccessScope(this);
-  }
-
-  public toString(): string {
-    return Unparser.unparse(this);
   }
 }
 
@@ -826,14 +439,6 @@ export class AccessMemberExpression {
     }
     return val;
   }
-
-  public accept<T>(visitor: IVisitor<T>): T {
-    return visitor.visitAccessMember(this);
-  }
-
-  public toString(): string {
-    return Unparser.unparse(this);
-  }
 }
 
 export class AccessKeyedExpression {
@@ -864,14 +469,6 @@ export class AccessKeyedExpression {
     const key = this.key.evaluate(s, e, null) as string;
     return instance[key] = val;
   }
-
-  public accept<T>(visitor: IVisitor<T>): T {
-    return visitor.visitAccessKeyed(this);
-  }
-
-  public toString(): string {
-    return Unparser.unparse(this);
-  }
 }
 
 export class CallScopeExpression {
@@ -901,14 +498,6 @@ export class CallScopeExpression {
 
   public assign(_s: Scope, _e: IAstEvaluator | null, _obj: unknown): unknown {
     return void 0;
-  }
-
-  public accept<T>(visitor: IVisitor<T>): T {
-    return visitor.visitCallScope(this);
-  }
-
-  public toString(): string {
-    return Unparser.unparse(this);
   }
 }
 
@@ -960,14 +549,6 @@ export class CallMemberExpression {
   public assign(_s: Scope, _e: IAstEvaluator | null, _obj: unknown): unknown {
     return void 0;
   }
-
-  public accept<T>(visitor: IVisitor<T>): T {
-    return visitor.visitCallMember(this);
-  }
-
-  public toString(): string {
-    return Unparser.unparse(this);
-  }
 }
 
 export class CallFunctionExpression {
@@ -997,14 +578,6 @@ export class CallFunctionExpression {
 
   public assign(_s: Scope, _e: IAstEvaluator | null, _obj: unknown): unknown {
     return void 0;
-  }
-
-  public accept<T>(visitor: IVisitor<T>): T {
-    return visitor.visitCallFunction(this);
-  }
-
-  public toString(): string {
-    return Unparser.unparse(this);
   }
 }
 
@@ -1103,14 +676,6 @@ export class BinaryExpression {
   public assign(_s: Scope, _e: IAstEvaluator | null, _obj: unknown): unknown {
     return void 0;
   }
-
-  public accept<T>(visitor: IVisitor<T>): T {
-    return visitor.visitBinary(this);
-  }
-
-  public toString(): string {
-    return Unparser.unparse(this);
-  }
 }
 
 export class UnaryExpression {
@@ -1146,14 +711,6 @@ export class UnaryExpression {
   public assign(_s: Scope, _e: IAstEvaluator | null, _obj: unknown): unknown {
     return void 0;
   }
-
-  public accept<T>(visitor: IVisitor<T>): T {
-    return visitor.visitUnary(this);
-  }
-
-  public toString(): string {
-    return Unparser.unparse(this);
-  }
 }
 export class PrimitiveLiteralExpression<TValue extends null | undefined | number | boolean | string = null | undefined | number | boolean | string> {
   public static readonly $undefined: PrimitiveLiteralExpression<undefined> = new PrimitiveLiteralExpression<undefined>(void 0);
@@ -1176,14 +733,6 @@ export class PrimitiveLiteralExpression<TValue extends null | undefined | number
   public assign(_s: Scope, _e: IAstEvaluator | null, _obj: unknown): unknown {
     return void 0;
   }
-
-  public accept<T>(visitor: IVisitor<T>): T {
-    return visitor.visitPrimitiveLiteral(this);
-  }
-
-  public toString(): string {
-    return Unparser.unparse(this);
-  }
 }
 
 export class ArrayLiteralExpression {
@@ -1202,14 +751,6 @@ export class ArrayLiteralExpression {
 
   public assign(_s: Scope, _e: IAstEvaluator | null, _obj: unknown): unknown {
     return void 0;
-  }
-
-  public accept<T>(visitor: IVisitor<T>): T {
-    return visitor.visitArrayLiteral(this);
-  }
-
-  public toString(): string {
-    return Unparser.unparse(this);
   }
 }
 
@@ -1235,14 +776,6 @@ export class ObjectLiteralExpression {
   public assign(_s: Scope, _e: IAstEvaluator | null, _obj: unknown): unknown {
     return void 0;
   }
-
-  public accept<T>(visitor: IVisitor<T>): T {
-    return visitor.visitObjectLiteral(this);
-  }
-
-  public toString(): string {
-    return Unparser.unparse(this);
-  }
 }
 
 export class TemplateExpression {
@@ -1267,14 +800,6 @@ export class TemplateExpression {
 
   public assign(_s: Scope, _e: IAstEvaluator | null, _obj: unknown): unknown {
     return void 0;
-  }
-
-  public accept<T>(visitor: IVisitor<T>): T {
-    return visitor.visitTemplate(this);
-  }
-
-  public toString(): string {
-    return Unparser.unparse(this);
   }
 }
 
@@ -1306,14 +831,6 @@ export class TaggedTemplateExpression {
 
   public assign(_s: Scope, _e: IAstEvaluator | null, _obj: unknown): unknown {
     return void 0;
-  }
-
-  public accept<T>(visitor: IVisitor<T>): T {
-    return visitor.visitTaggedTemplate(this);
-  }
-
-  public toString(): string {
-    return Unparser.unparse(this);
   }
 }
 
@@ -1347,14 +864,6 @@ export class ArrayBindingPattern {
     // TODO
     return void 0;
   }
-
-  public accept<T>(visitor: IVisitor<T>): T {
-    return visitor.visitArrayBindingPattern(this);
-  }
-
-  public toString(): string {
-    return Unparser.unparse(this);
-  }
 }
 
 export class ObjectBindingPattern {
@@ -1380,14 +889,6 @@ export class ObjectBindingPattern {
     // TODO
     return void 0;
   }
-
-  public accept<T>(visitor: IVisitor<T>): T {
-    return visitor.visitObjectBindingPattern(this);
-  }
-
-  public toString(): string {
-    return Unparser.unparse(this);
-  }
 }
 
 export class BindingIdentifier {
@@ -1401,14 +902,6 @@ export class BindingIdentifier {
 
   public evaluate(_s: Scope, _e: IAstEvaluator | null, _c: IConnectable | null): string {
     return this.name;
-  }
-
-  public accept<T>(visitor: IVisitor<T>): T {
-    return visitor.visitBindingIdentifier(this);
-  }
-
-  public toString(): string {
-    return Unparser.unparse(this);
   }
 }
 
@@ -1442,14 +935,6 @@ export class ForOfStatement {
     if (this.iterable.hasUnbind) {
       this.iterable.unbind(s, b);
     }
-  }
-
-  public accept<T>(visitor: IVisitor<T>): T {
-    return visitor.visitForOfStatement(this);
-  }
-
-  public toString(): string {
-    return Unparser.unparse(this);
   }
 }
 
@@ -1489,14 +974,6 @@ export class Interpolation {
 
   public assign(_s: Scope, _e: IAstEvaluator | null, _obj: unknown): unknown {
     return void 0;
-  }
-
-  public accept<T>(visitor: IVisitor<T>): T {
-    return visitor.visitInterpolation(this);
-  }
-
-  public toString(): string {
-    return Unparser.unparse(this);
   }
 }
 
@@ -1546,14 +1023,6 @@ export class DestructuringAssignmentExpression {
       }
     }
   }
-
-  public accept<T>(visitor: IVisitor<T>): T {
-    return visitor.visitDestructuringAssignmentExpression(this);
-  }
-
-  public toString(): string {
-    return Unparser.unparse(this);
-  }
 }
 
 /** This is an internal API */
@@ -1583,14 +1052,6 @@ export class DestructuringAssignmentSingleExpression {
       source = this.initializer?.evaluate(s, l, null);
     }
     this.target.assign(s, l, source);
-  }
-
-  public accept<T>(visitor: IVisitor<T>): T {
-    return visitor.visitDestructuringAssignmentSingleExpression(this);
-  }
-
-  public toString(): string {
-    return Unparser.unparse(this);
   }
 }
 
@@ -1639,14 +1100,6 @@ export class DestructuringAssignmentRestExpression {
     }
     this.target.assign(s, l, restValue);
   }
-
-  public accept<T>(_visitor: IVisitor<T>): T {
-    return _visitor.visitDestructuringAssignmentRestExpression(this);
-  }
-
-  public toString(): string {
-    return Unparser.unparse(this);
-  }
 }
 
 export class ArrowFunction {
@@ -1681,14 +1134,6 @@ export class ArrowFunction {
 
   public assign(_s: Scope, _e: IAstEvaluator | null, _value: unknown): void {
     return void 0;
-  }
-
-  public accept<T>(visitor: IVisitor<T>): T {
-    return visitor.visitArrowFunction(this);
-  }
-
-  public toString(): string {
-    return Unparser.unparse(this);
   }
 }
